+++ packages/datepicker/CHANGELOG.md
@@ -3,6 +3,19 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.9.0](https://github.com/inavac182/ui-react/compare/@uireact/datepicker@0.7.2...@uireact/datepicker@0.9.0) (2024-05-02)
+
+
+### Features
+
+* hot load file changes in packages and update app docs ([4343546](https://github.com/inavac182/ui-react/commit/4343546a7739f011875050723426f29231d561a8))
+* improve docs texts ([d871b93](https://github.com/inavac182/ui-react/commit/d871b93ba8bcacc81020325c1f8aa9ef63a26c60))
+* **UiMenu:** :adding framer motion ([#320](https://github.com/inavac182/ui-react/issues/320)) ([cc158a6](https://github.com/inavac182/ui-react/commit/cc158a6741aee5fd6ab4320989b0b80c6618021d))
+
+
+
+
+
 # [0.8.0](https://github.com/inavac182/ui-react/compare/@uireact/datepicker@0.7.2...@uireact/datepicker@0.8.0) (2024-04-30)